Move more from card to contract
Change-type: patch
