+++ lib/contracts/index.ts
