+++ test/integration/utils.spec.ts
@@ -14,7 +14,7 @@ afterAll(() => {
 
 describe('.hasCard()', () => {
 	test('id = yes (exists), slug = yes (exists)', async () => {
-		const card = await ctx.kernel.insertCard(ctx.logContext, ctx.session, {
+		const card = await ctx.kernel.insertContract(ctx.logContext, ctx.session, {
 			slug: coreTestUtils.generateRandomSlug(),
 			type: 'card@1.0.0',
 			version: '1.0.0',
@@ -30,7 +30,7 @@ describe('.hasCard()', () => {
 	});
 
 	test('id = yes (exists), slug = yes (not exist)', async () => {
-		const card = await ctx.kernel.insertCard(ctx.logContext, ctx.session, {
+		const card = await ctx.kernel.insertContract(ctx.logContext, ctx.session, {
 			slug: coreTestUtils.generateRandomSlug(),
 			type: 'card@1.0.0',
 			version: '1.0.0',
@@ -46,7 +46,7 @@ describe('.hasCard()', () => {
 	});
 
 	test('id = yes (not exist), slug = yes (exists)', async () => {
-		const card = await ctx.kernel.insertCard(ctx.logContext, ctx.session, {
+		const card = await ctx.kernel.insertContract(ctx.logContext, ctx.session, {
 			slug: coreTestUtils.generateRandomSlug(),
 			type: 'card@1.0.0',
 			version: '1.0.0',
@@ -72,7 +72,7 @@ describe('.hasCard()', () => {
 	});
 
 	test('id = no, slug = yes (exists)', async () => {
-		const card = await ctx.kernel.insertCard(ctx.logContext, ctx.session, {
+		const card = await ctx.kernel.insertContract(ctx.logContext, ctx.session, {
 			slug: coreTestUtils.generateRandomSlug(),
 			type: 'card@1.0.0',
 			version: '1.0.0',
